remove provider/add ./update test that wasnt useful
